Security: parameterize/validate SQL injection sinks
Phase 2a batch 2 — fix the raw-SQL injection points from the audit.

- logs.php (message search): bind the five LIKE clauses instead of
  interpolating the `q` term into whereRaw (was authenticated SQLi).
- logs.php (list + radius) and reports.php (by-date/activation): cast the
  `keep` retention value to int before "INTERVAL $keep DAY" in the DELETE
  statements against tbl_logs / radpostauth / tbl_transactions.
- customers.php: normalise `orderby` to exactly asc|desc, closing ORDER BY
  injection via order_by_expr on the lastname sort.
- settings.php (dbrestore): validate each restored table name as a plain
  identifier and backtick-quote it before TRUNCATE, blocking stacked-statement
  injection through a crafted backup file.

Behaviour is unchanged for legitimate inputs.

+++ system/controllers/reports.php
@@ -259,7 +259,7 @@ switch ($action) {
     case 'by-date':
     case 'activation':
         $q = (_post('q') ? _post('q') : _get('q'));
-        $keep = _post('keep');
+        $keep = (int) _post('keep');
         if (!empty($keep)) {
             ORM::raw_execute("DELETE FROM tbl_transactions WHERE date < UNIX_TIMESTAMP(DATE_SUB(NOW(), INTERVAL $keep DAY))");
             r2(getUrl('logs/list/'), 's', "Delete logs older than $keep days");
